consolidate field property state
